+++ pkg/cmd/release.go
@@ -12,6 +12,8 @@ import (
 )
 
 var fileName string
+var prefix string
+var postfix string
 
 // ddCmd represents the dd command
 var release = &cobra.Command{
@@ -21,6 +23,10 @@ var release = &cobra.Command{
 	PreRun:  core.ToggleDebug,
 	Example: releaseReleaseDisc,
 	Run: func(cmd *cobra.Command, args []string) {
+		if !core.ValidateIsDIrectory(o.Output) {
+			log.Errorf("The Set output is not a valid directory: %s", o.Output)
+		}
+
 		option := gits.FlagsOptions{GitAuth: o.GitAuth, GitBranch: o.GitBranch, GitUser: o.GitUser, GitEmail: o.GitEmail, GitKeyPath: o.GitKeyPath, Output: o.Output, CommitHash: o.CommitHash, Orgenization: o.Orgenization, Pat: o.Pat, Project: o.Project, RepoPath: o.RepoPath, DryRun: o.DryRun, Gitpush: o.Gitpush}
 		r, commentsArray, newVersionTag, latestTag := option.MainGits()
 		sortingForMD, workitemsID := markdown.SortCommitsForMD(commentsArray, option.Orgenization, option.Project, option.Pat)
@@ -37,12 +43,18 @@ var release = &cobra.Command{
 		if setBool || option.DryRun {
 			var output string
 			if fileName == "" {
-				output = filepath.Join(option.Output, newVersionTag)
+				if prefix != "" {
+					output = filepath.Join(option.Output, prefix+newVersionTag)
+				} else if postfix != "" {
+					output = filepath.Join(option.Output, newVersionTag+postfix)
+				} else {
+					output = filepath.Join(option.Output, newVersionTag)
+				}
 			} else {
 				output = filepath.Join(option.Output, fileName)
 			}
 			file := markdown.HasMDSuffix(output, "md")
-			log.Info(file)
+			log.Warningf("The file will be saved at this path: %s", file)
 			markdown.WriteToMD(sortingForMD, latestTag, newVersionTag, file)
 		}
 	},
@@ -61,9 +73,14 @@ func init() {
 	release.Flags().StringVar(&o.GitEmail, "git-email", ".", "Set email to tag with")
 	release.Flags().StringVar(&o.GitKeyPath, "git-keyPath", "~/.ssh/id_rsa", "Set email to tag with")
 	release.Flags().StringVar(&o.GitAuth, "auth", "https", "Set Auth type (ssh or https")
+	release.Flags().StringVar(&prefix, "prefix", "", "add a costume prefix to the version name. for example: <prefix->1.0.0.md")
+	release.Flags().StringVar(&postfix, "postfix", "", "add a costume postfix to the version name. for example: 1.0.0<-postfix>.md")
 	release.Flags().BoolVar(&o.DryRun, "dry-run", false, "If true, only run a dry-run with cli output")
 	release.Flags().BoolVar(&o.Gitpush, "git-push", false, "If true, only run a dry-run with cli output")
 	release.MarkFlagsRequiredTogether("org", "project", "pat")
+	release.MarkFlagsMutuallyExclusive("postfix", "prefix")
+	release.MarkFlagsMutuallyExclusive("filename", "prefix")
+	release.MarkFlagsMutuallyExclusive("filename", "postfix")
 	release.MarkFlagRequired("hash")
 	release.MarkFlagRequired("repo-path")
 	rootCmd.AddCommand(release)