+++ .github/workflows/CHANGELOG.md
@@ -1,5 +1,8 @@
 # gitops-release-manager
 ## Fix
-- add support to commits not working in convention
+
 ## Changes
+- add support to prefix and postfix for filename
+- added directory validation in start of run.
 ## Breaking changes
+-- filename flag cant be set with prefix or postfix